requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
platformdirs==4.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
